created DMRGWorker class for two-site DMRG
