+++ itensor/dmrg.h
@@ -3,6 +3,8 @@
 #include "mpo.h"
 #include "sparse.h"
 #include "davidson.h"
+#include "Sweeps.h"
+#include "DMRGOpts.h"
 
 template<class Tensor,class TensorSet, class OpTensorSet>
 void applyProjOp(const Tensor& phi, const TensorSet& L, const TensorSet& R, const OpTensorSet& H, Tensor& Hphi)
@@ -208,180 +210,6 @@ Vector doDavidson(std::vector<Tensor>& phi, const TensorSet& mpoh, const TensorS
         energies(1+n) = evals(1+n);
     }
     return energies;
-}
-
-enum SweepScheme {ramp_m, fixed_m, fixed_cutoff};
-
-inline void sweepnext(int &l, int &ha, int N, int min_l = 1)
-{
-    if(ha == 1)
-	{
-        if(++l == N) 
-            l = N-1, ha = 2;
-        return;
-	}
-    if(l-- == min_l) ha = 3;
-}
-
-class Sweeps
-{
-public:
-    SweepScheme scheme;
-    int Minm;
-    std::vector<int>  Maxm, Niter;
-    std::vector<Real> Cutoff;
-    int Nsweep;
-    int num_site_center;		// May not be implemented in some cases
-    Sweeps(SweepScheme sch, int nsw, int _minm, int _maxm, Real _cut)
-	    : scheme(sch), Minm(_minm), Maxm(nsw+1), Niter(nsw+1,4), Cutoff(nsw+1), Nsweep(nsw), num_site_center(2)
-	{
-        if(scheme == ramp_m)
-        {
-            for(int s = 1; s <= Nsweep; s++)
-            { Cutoff.at(s) = _cut; Maxm.at(s) = (int)(_minm + (s-1.0)/nsw * (_maxm - _minm)); }
-        }
-        else if(scheme == fixed_m || scheme == fixed_cutoff)
-        {
-            for(int s = 1; s <= Nsweep; s++)
-            { Cutoff.at(s) = _cut; Maxm.at(s) = _maxm; }
-        }
-
-        for(int s = 1; s <= min(Nsweep,4); s++)
-        { Niter.at(s) = 10 - s; }
-	}
-    Real cutoff(int sw) const { return Cutoff.at(sw); }
-    int minm(int sw) const { return Minm; }
-    int maxm(int sw) const { return Maxm.at(sw); }
-    int nsweep() const { return Nsweep; }
-    int niter(int sw) const { return Niter.at(sw); }
-};
-
-//Class for fine-tuning DMRG algorithms
-class DMRGOpts
-{
-private:
-    Vector center_eigs;
-    Real energy_errgoal;    //Stop DMRG once energy has converged to this precision
-    Real orth_weight;       //How much to penalize non-orthogonality in multiple-state DMRG
-    bool printeigs;         //Print slowest decaying eigenvalues after every sweep
-public:
-    bool quiet;             //Show/don't show info after every step
-
-    Real energyErrgoal() const { return energy_errgoal; }
-    void energyErrgoal(Real val) { energy_errgoal = val; }
-
-    Real orthWeight() const { return orth_weight; }
-    void orthWeight(Real val) { orth_weight = val; }
-
-    bool printEigs() const { return printeigs; }
-    void printEigs(bool val) { printeigs = val; }
-
-    DMRGOpts() : 
-    energy_errgoal(-1), 
-    orth_weight(1),
-    printeigs(true), 
-    quiet(true)
-    { }
-
-    virtual void measure(int sw, int ha, int b, const SVDWorker& svd, Real energy)
-    {
-        if(printeigs)
-        {
-            if(b == 1 && ha == 2) 
-            {
-                std::cout << "\n    Largest m during sweep " << sw << " was " << svd.maxEigsKept() << "\n";
-                std::cout << "    Largest truncation error: " << svd.maxTruncerr() << "\n";
-                Vector center_eigs = svd.eigsKept(svd.NN()/2);
-                std::cout << "    Eigs at center bond: ";
-                for(int j = 1; j <= min(center_eigs.Length(),10); ++j) 
-                {
-                    std::cout << boost::format(center_eigs(j) > 1E-2 ? ("%.2f") : ("%.2E")) % center_eigs(j);
-                    std::cout << ((j != min(center_eigs.Length(),10)) ? ", " : "\n");
-                }
-                std::cout << boost::format("    Energy after sweep %d is %f\n") % sw % energy;
-            }
-        }
-    }
-
-    virtual bool checkDone(int sw, Real energy)
-    {
-        static Real last_energy;
-
-        if(sw == 1) last_energy = 1000;
-        if(energy_errgoal > 0 && sw%2 == 0)
-        {
-            Real dE = fabs(energy-last_energy);
-            if(dE < energy_errgoal)
-            {
-                std::cout << boost::format("    Energy error goal met (dE = %E); returning after %d sweeps.\n") % dE % sw;
-                return true;
-            }
-        }
-        last_energy = energy;
-
-        return false;
-    }
-};
-
-template <class Tensor>
-Real dmrg(MPSt<Tensor>& psi, const MPOt<Tensor>& H, const Sweeps& sweeps, DMRGOpts& opts)
-{
-    const Real orig_cutoff = psi.cutoff(); 
-    const int orig_minm = psi.minm(), orig_maxm = psi.maxm();
-    int debuglevel = (opts.quiet ? 0 : 1);
-    int N = psi.NN();
-    Real energy = 0;
-
-    psi.position(1);
-    //if(H.is_complex()) psi.AAnc(1) *= Complex_1;
-
-    std::vector<Tensor> PH(N+1);
-    for(int l = N-1; l >= 2; --l) psi.projectOp(l+1,Fromright,PH[l+1],H.AA(l+1),PH[l]);
-
-    for(int sw = 1; sw <= sweeps.nsweep(); ++sw)
-    {
-        psi.cutoff(sweeps.cutoff(sw)); psi.minm(sweeps.minm(sw)); psi.maxm(sweeps.maxm(sw));
-        for(int b = 1, ha = 1; ha != 3; sweepnext(b,ha,N))
-        {
-            if(!opts.quiet) std::cout << boost::format("Sweep=%d, HS=%d, Bond=(%d,%d)\n") % sw % ha % b % (b+1);
-
-            energy = psi.bondDavidson(b,H.bondTensor(b),PH[b],PH[b+1],
-                     sweeps.niter(sw),debuglevel,(ha==1?Fromleft:Fromright));
-
-            if(!opts.quiet) 
-            { 
-                std::cout << boost::format("    Truncated to Cutoff=%.1E, Min_m=%d, Max_m=%d\n") 
-                                      % sweeps.cutoff(sw) % sweeps.minm(sw) % sweeps.maxm(sw);
-                std::cout << boost::format("    Trunc. err=%.1E, States kept=%s\n")
-                                      % psi.svd().truncerr(b) % psi.LinkInd(b).showm();
-            }
-
-            opts.measure(sw,ha,b,psi.svd(),energy);
-
-            if(ha == 1 && b != N-1) psi.projectOp(b,Fromleft,PH[b],H.AA(b),PH[b+1]);
-            if(ha == 2 && b != 1)   psi.projectOp(b+1,Fromright,PH[b+1],H.AA(b+1),PH[b]);
-        } //for loop over b
-
-        if(opts.checkDone(sw,energy))
-        {
-            psi.cutoff(orig_cutoff); 
-            psi.minm(orig_minm); 
-            psi.maxm(orig_maxm);
-            return energy;
-        }
-
-    } //for loop over sw
-
-    psi.cutoff(orig_cutoff); 
-    psi.minm(orig_minm); 
-    psi.maxm(orig_maxm);
-    return energy;
-}
-template <class MPSType, class MPOType>
-Real dmrg(MPSType& psi, const MPOType& H, const Sweeps& sweeps)
-{
-    DMRGOpts opts; 
-    return dmrg(psi,H,sweeps,opts);
 }
 
 void onesite_sweepnext(int &l, int &ha, int N)
